pyproject.toml
[project]
name = "wikinator"
version = "0.5.0"
description = "Convert Google docs to markdown"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.175.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gql[all]>=3.5.3",
    "pytest>=8.4.1",
    "python-docx>=1.2.0",
    "typer-slim>=0.16.0",
]

[project.urls]
Documentation = "https://github.com/philion/wikinator#readme"
Issues = "https://github.com/philion/wikinator/issues"
Source = "https://github.com/philion/wikinator"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.coverage.run]
source_pkgs = ["wikinator", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
clw = ["src/wikinator", "*/wikinator/src/wikinator"]
tests = ["tests", "*/wikinator/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.hatch]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wikinator"]

[project.scripts]
wikinator = "wikinator.app:main"
